Check for trailing ':' before iterating over canindent array
We need to do it anyway, if we match, so do it once before the loop
(we can also do it after the loop, but I am going to remove this loop
soon anyway).

This speeds up entire dmt by 20%. I.e. on very large files (200MB),
gdc -O3 -frelease compiled dmt, dmt finishes in 3.95s, instead of 5.16s